+++ core/src/main/java/zemeckis/MacroTaskExecutor.java
@@ -1,14 +1,33 @@
 package zemeckis;
 
+import elemental2.dom.MessageChannel;
+
 /**
  * Run tasks in next MacroTask.
  */
 final class MacroTaskExecutor
   extends RoundBasedExecutor
 {
+  private final MessageChannel _channel = Zemeckis.useMessageChannelToScheduleTasks() ? new MessageChannel() : null;
+
+  MacroTaskExecutor()
+  {
+    if ( Zemeckis.useMessageChannelToScheduleTasks() )
+    {
+      _channel.port1.onmessage = m -> activate();
+    }
+  }
+
   @Override
   void scheduleForActivation()
   {
-    TemporalScheduler.delayedTask( Zemeckis.areNamesEnabled() ? "MacroTaskExecutor" : null, this::activate, 0 );
+    if ( Zemeckis.useMessageChannelToScheduleTasks() )
+    {
+      _channel.port2.postMessage( null );
+    }
+    else
+    {
+      TemporalScheduler.delayedTask( Zemeckis.areNamesEnabled() ? "MacroTaskExecutor" : null, this::activate, 0 );
+    }
   }
 }