+++ core/src/main/java/zemeckis/ZemeckisConfig.java
@@ -16,6 +16,7 @@ final class ZemeckisConfig
   private static boolean ENABLE_NAMES = PROVIDER.areNamesEnabled();
   private static boolean PURGE_ON_RUNAWAY = PROVIDER.purgeTasksWhenRunawayDetected();
   private static boolean UNCAUGHT_ERROR_HANDLERS = PROVIDER.areUncaughtErrorHandlersEnabled();
+  private static boolean USE_MESSAGE_CHANNEL_TO_SCHEDULE_TASKS = PROVIDER.useMessageChannelToScheduleTasks();
   private static final String LOGGER_TYPE = PROVIDER.loggerType();
   private static final boolean JVM = PROVIDER.isJvm();
 
@@ -51,6 +52,11 @@ final class ZemeckisConfig
   static boolean areUncaughtErrorHandlersEnabled()
   {
     return UNCAUGHT_ERROR_HANDLERS;
+  }
+
+  static boolean useMessageChannelToScheduleTasks()
+  {
+    return USE_MESSAGE_CHANNEL_TO_SCHEDULE_TASKS;
   }
 
   static String loggerType()
@@ -92,6 +98,13 @@ final class ZemeckisConfig
 
     @GwtIncompatible
     @Override
+    boolean useMessageChannelToScheduleTasks()
+    {
+      return "true".equals( System.getProperty( "zemeckis.use_message_channel_to_schedule_tasks", "true" ) );
+    }
+
+    @GwtIncompatible
+    @Override
     String loggerType()
     {
       return System.getProperty( "zemeckis.logger", PRODUCTION_MODE ? "basic" : "proxy" );
@@ -126,6 +139,11 @@ final class ZemeckisConfig
     boolean areUncaughtErrorHandlersEnabled()
     {
       return "true" == System.getProperty( "zemeckis.enable_uncaught_error_handlers" );
+    }
+
+    boolean useMessageChannelToScheduleTasks()
+    {
+      return "true" == System.getProperty( "zemeckis.use_message_channel_to_schedule_tasks" );
     }
 
     String loggerType()