+++ core/src/main/java/zemeckis/TemporalScheduler.java
@@ -1,12 +1,21 @@
 package zemeckis;
 
+import elemental2.dom.Blob;
 import elemental2.dom.DomGlobal;
+import elemental2.dom.MessageEvent;
+import elemental2.dom.URL;
+import elemental2.dom.Worker;
+import grim.annotations.OmitSymbol;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
 import org.jetbrains.annotations.TestOnly;
 import static org.realityforge.braincheck.Guards.*;
 
@@ -116,7 +125,69 @@ final class TemporalScheduler
 
   private static abstract class AbstractScheduler
   {
+    private static final boolean ENABLE_WORKERS = Zemeckis.useWorkerToScheduleDelayedTasks();
+    @Nonnull
+    private final String SRC =
+      "var timers = {};\n" +
+      "\n" +
+      "function cancelTimer(id) {\n" +
+      "  let timerId = timers[id] || 0;\n" +
+      "  self.clearTimeout(timerId);\n" +
+      "  timers.delete(id);\n" +
+      "}\n" +
+      "\n" +
+      "function createTimer(id, delay) {\n" +
+      "  timers[id] = self.setTimeout(() => {\n" +
+      "    self.postMessage({ type: 'dt', id: id });\n" +
+      "  }, delay);\n" +
+      "}\n" +
+      "\n" +
+      "function cancelPeriodicTimer(id) {\n" +
+      "  let timerId = timers[id] || 0;\n" +
+      "  self.clearInterval(timerId);\n" +
+      "  timers.delete(id);\n" +
+      "}\n" +
+      "\n" +
+      "function createPeriodicTimer(id, period) {\n" +
+      "  timers[id] = self.setInterval(() => {\n" +
+      "    self.postMessage({ type: 'pt', id: id });\n" +
+      "  }, period);\n" +
+      "}\n" +
+      "\n" +
+      "self.onmessage = m => {\n" +
+      "  if (m.data && m.data.action === '+' && m.data.type === 'dt') {\n" +
+      "    createTimer(m.data.id, m.data.delay);\n" +
+      "  } else if (m.data && m.data.action === '-' && m.data.type === 'dt') {\n" +
+      "    cancelTimer(m.data.id);\n" +
+      "  } else if (m.data && m.data.action === '+' && m.data.type === 'pt') {\n" +
+      "    createPeriodicTimer(m.data.id, m.data.period);\n" +
+      "  } else if (m.data && m.data.action === '-' && m.data.type === 'pt') {\n" +
+      "    cancelPeriodicTimer(m.data.id);\n" +
+      "  }\n" +
+      "};";
     private final long _schedulerStart = System.currentTimeMillis();
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    private final Worker _worker =
+      ENABLE_WORKERS ?
+      new Worker(
+        URL.createObjectURL(
+          new Blob(
+            new Blob.ConstructorBlobPartsArrayUnionType[]{ Blob.ConstructorBlobPartsArrayUnionType.of( SRC ) }
+          )
+        )
+      ) :
+      null;
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    private final Map<Double, Runnable> _workerTasks = ENABLE_WORKERS ? new HashMap<>() : null;
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    private double _nextTimerId = 1;
+
+    {
+      if ( ENABLE_WORKERS )
+      {
+        _worker.onmessage = this::onWorkerMessage;
+      }
+    }
 
     final long getSchedulerStart()
     {
@@ -125,6 +196,10 @@ final class TemporalScheduler
 
     void shutdown()
     {
+      if ( Zemeckis.useWorkerToScheduleDelayedTasks() )
+      {
+        _worker.terminate();
+      }
     }
 
     final int now()
@@ -146,8 +221,23 @@ final class TemporalScheduler
     @Nonnull
     Cancelable doDelayedTask( @Nonnull final Runnable task, final int delay )
     {
-      final double timeoutId = DomGlobal.setTimeout( v -> task.run(), delay );
-      return () -> DomGlobal.clearTimeout( timeoutId );
+      if ( Zemeckis.useWorkerToScheduleDelayedTasks() )
+      {
+        final double id = _nextTimerId++;
+        _workerTasks.put( id, task );
+        final JsPropertyMap<Object> message = msg( "+", "pt", id );
+        message.set( "delay", delay );
+        _worker.postMessage( message );
+        return () -> {
+          _workerTasks.remove( id );
+          _worker.postMessage( msg( "-", "dt", id ) );
+        };
+      }
+      else
+      {
+        final double timeoutId = DomGlobal.setTimeout( v -> task.run(), delay );
+        return () -> DomGlobal.clearTimeout( timeoutId );
+      }
     }
 
     @Nonnull
@@ -166,8 +256,58 @@ final class TemporalScheduler
     @Nonnull
     Cancelable doPeriodicTask( @Nonnull final Runnable task, final int period )
     {
-      final double timeoutId = DomGlobal.setInterval( v -> task.run(), period );
-      return () -> DomGlobal.clearInterval( timeoutId );
+      if ( Zemeckis.useWorkerToScheduleDelayedTasks() )
+      {
+        final double id = _nextTimerId++;
+        _workerTasks.put( id, task );
+        final JsPropertyMap<Object> message = msg( "+", "pt", id );
+        message.set( "period", period );
+        _worker.postMessage( message );
+        return () -> {
+          _workerTasks.remove( id );
+          _worker.postMessage( msg( "-", "pt", id ) );
+        };
+      }
+      else
+      {
+        final double timeoutId = DomGlobal.setInterval( v -> task.run(), period );
+        return () -> DomGlobal.clearInterval( timeoutId );
+      }
+    }
+
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    private void onWorkerMessage( @Nonnull final MessageEvent<Object> event )
+    {
+      if ( null != event.data )
+      {
+        final JsPropertyMap<Object> data = Js.asPropertyMap( event.data );
+        final double id = data.getAsAny( "id" ).asDouble();
+        final String type = data.getAsAny( "type" ).asString();
+        if ( "dt".equals( type ) )
+        {
+          runTaskIfPresent( _workerTasks.remove( id ) );
+        }
+        else if ( "pt".equals( type ) )
+        {
+          runTaskIfPresent( _workerTasks.get( id ) );
+        }
+      }
+    }
+
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    @Nonnull
+    private JsPropertyMap<Object> msg( @Nonnull final String action, @Nonnull final String type, final double id )
+    {
+      return JsPropertyMap.of( "action", action, "type", type, "id", id );
+    }
+
+    @OmitSymbol( unless = "zemeckis.use_worker_to_schedule_delayed_tasks" )
+    private void runTaskIfPresent( @Nullable final Runnable task )
+    {
+      if ( null != task )
+      {
+        task.run();
+      }
     }
   }
 }