+++ core/src/test/java/zemeckis/test/JDependTest.java
@@ -27,10 +27,12 @@ public class JDependTest
     final JavaPackage zemeckis = constraint.addPackage( "zemeckis" );
     final JavaPackage braincheck = constraint.addPackage( "org.realityforge.braincheck" );
     final JavaPackage jsinterop = constraint.addPackage( "jsinterop.annotations" );
+    final JavaPackage jsinteropBase = constraint.addPackage( "jsinterop.base" );
     final JavaPackage elementalPromise = constraint.addPackage( "elemental2.promise" );
     final JavaPackage elementalDom = constraint.addPackage( "elemental2.dom" );
 
     zemeckis.dependsUpon( jsinterop );
+    zemeckis.dependsUpon( jsinteropBase );
     zemeckis.dependsUpon( braincheck );
     zemeckis.dependsUpon( elementalPromise );
     zemeckis.dependsUpon( elementalDom );